수정: steppingPositions 함수로 밟는 위치 반환하도록 수정
